[CL-1576] Extract event and page change tracking logic into actions file
